pyproject.toml
[tool.poetry]
name = "gunicorn-reload"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "uwsgi_reload"}]

[tool.poetry.dependencies]
python = "^3.8"
django = "<4.0.0"
gunicorn = "20.1.0"
djangorestframework = "3.11.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
